requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
blessings==1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
climage==0.1.3
colorama==0.4.4
cryptography==36.0.1
curtsies==0.3.10
cwcwidth==0.1.6
discord-py-slash-command==3.0.3
discord.py==1.7.3
enquiries==0.1.0
google==3.0.0
h11==0.13.0
idna==3.3
joblib==1.1.0
kdtree==0.16
keyboard==0.13.5
lxml==4.7.1
multidict==5.2.0
nltk==3.6.7
outcome==1.1.0
pick==1.2.0
Pillow==9.0.0
pycparser==2.21
pyOpenSSL==21.0.0
regex==2021.11.10
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
terminal-img==0.0.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
urllib3==1.26.7
wsproto==1.0.0
yarl==1.7.2

pymongo~=4.0.1
opencv-python~=4.5.5.62
numpy~=1.22.1
matplotlib~=3.5.1
chess~=1.8.0